pyproject.toml
[tool.poetry]
name = "licencjat"
version = "0.1.0"
description = ""
authors = ["Kamil Gierada <vlaska8888@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.8,<3.10"
numpy = "^1.20.2"
Cython = "^0.29.22"
Pillow = "^8.2.0"
matplotlib = "^3.4.1"
imageio = "^2.9.0"
pygame = "^2.0.1"
numba = "^0.53.1"
PyYAML = "^5.4.1"

[tool.poetry.dev-dependencies]
pylint = "^2.7.4"
mypy = "^0.812"
autopep8 = "^1.5.6"
ipython = "^7.22.0"
jedi = "0.17.2"
pylama = "^7.7.1"
pytest = "^6.2.3"
loguru = "^0.5.3"
rope = "^0.19.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.autopep8]
max_line_length = 79

[tool.mypy]
warn_no_return = false
